### 2022-11-09 #4920 @editorialbot
[PRE REVIEW]: DeBIR: A Python Package for Dense Bi-Encoder Information Retrieval
**Submitting author:** @ayuei (<a href="http://orcid.org/0000-0003-1787-8090">Vincent Nguyen</a>)
**Repository:** https://github.com/Ayuei/DeBEIR
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.0.1
