### 2022-11-11 d034a99
Delete assets directory
