+++ assets/qrels-clinical_trials.txt
@@ -1,3870 +1,0 @@
-20141	0	NCT00000408	0
-20141	0	NCT00000492	1
-20141	0	NCT00000501	0
-20141	0	NCT00001853	0
-20141	0	NCT00004727	0
-20141	0	NCT00005127	2
-20141	0	NCT00005485	2
-20141	0	NCT00005757	2
-20141	0	NCT00077948	0
-20141	0	NCT00108381	0
-20141	0	NCT00124969	0
-20141	0	NCT00129233	1
-20141	0	NCT00134160	0
-20141	0	NCT00141583	1
-20141	0	NCT00143195	2
-20141	0	NCT00149227	1
-20141	0	NCT00162344	1
-20141	0	NCT00166231	0
-20141	0	NCT00175279	2
-20141	0	NCT00185120	0
-20141	0	NCT00190489	0
-20141	0	NCT00202566	0
-20141	0	NCT00225355	1
-20141	0	NCT00252421	0
-20141	0	NCT00285649	0
-20141	0	NCT00337116	0
-20141	0	NCT00348803	0
-20141	0	NCT00356707	0
-20141	0	NCT00373828	2
-20141	0	NCT00375089	0
-20141	0	NCT00413712	0
-20141	0	NCT00418392	0
-20141	0	NCT00424021	0
-20141	0	NCT00454662	0
-20141	0	NCT00462241	0
-20141	0	NCT00479908	1
-20141	0	NCT00486954	0
-20141	0	NCT00536224	0
-20141	0	NCT00598728	0
-20141	0	NCT00623454	0
-20141	0	NCT00624273	0
-20141	0	NCT00644319	0
-20141	0	NCT00683813	2
-20141	0	NCT00705952	0
-20141	0	NCT00736333	0
-20141	0	NCT00745043	0
-20141	0	NCT00752674	0
-20141	0	NCT00780260	0
-20141	0	NCT00797953	1
-20141	0	NCT00802230	1
-20141	0	NCT00808652	1
-20141	0	NCT00809029	2
-20141	0	NCT00832442	2
-20141	0	NCT00840593	0
-20141	0	NCT00848250	0
-20141	0	NCT00883987	0
-20141	0	NCT00896285	0
-20141	0	NCT00925119	0
-20141	0	NCT00932867	0
-20141	0	NCT00937365	0
-20141	0	NCT00951808	0
-20141	0	NCT00952744	2
-20141	0	NCT00974116	0
-20141	0	NCT00979199	1
-20141	0	NCT00986180	0
-20141	0	NCT00992927	0
-20141	0	NCT01028534	0
-20141	0	NCT01056367	0
-20141	0	NCT01080274	0
-20141	0	NCT01087723	0
-20141	0	NCT01134328	0
-20141	0	NCT01162902	1
-20141	0	NCT01167205	0
-20141	0	NCT01171911	0
-20141	0	NCT01175889	0
-20141	0	NCT01203696	1
-20141	0	NCT01211171	0
-20141	0	NCT01216020	0
-20141	0	NCT01228123	0
-20141	0	NCT01231165	1
-20141	0	NCT01250262	0
-20141	0	NCT01253486	2
-20141	0	NCT01276548	0
-20141	0	NCT01284179	0
-20141	0	NCT01290770	0
-20141	0	NCT01293019	0
-20141	0	NCT01332188	0
-20141	0	NCT01334203	0
-20141	0	NCT01365286	0
-20141	0	NCT01397994	1
-20141	0	NCT01407146	1
-20141	0	NCT01429194	1
-20141	0	NCT01444833	0
-20141	0	NCT01450020	0
-20141	0	NCT01484912	0
[…]
