### 2022-11-11 95c687c
Refactoring Sentence Transformer Dataset objects for more usability.
