### 2022-11-11 c009692
Fixed Parser init logic
