### 2022-12-08 #4920 @Ayuei
> @Ayuei – could you take a look a [this list of potential reviewers](https://bit.ly/joss-reviewers) and identify a few people who would be good candidates to review this submission?

Apologies for the late response. But good candidates would be those with a background in information retrieval and natural language processing. The primary programming language used in the package is Python.

I think the following are good candidates (in no particular order):
amitkumarj441
brunaw
elektrobohemian
JacksonMaxfield
KonradHoeffner
shruti-singh
