### 2022-12-09 #4920 @arfon
:wave: @KonradHoeffner @brunaw @elektrobohemian would any of you be willing to review this submission for JOSS? The submission under consideration is [DeBIR: A Python Package for Dense Bi-Encoder Information Retrieval](https://github.com/openjournals/joss-reviews/issues/4920), a Python package for information retrieval and natural language processing.

Many thanks
Arfon
