### 2022-12-15 768ac87
Fixed instructions for dev environment
Virtualenv instructions and pip installation.
