### 2022-12-15 4fdee02
bugfix: renamed test set extraction and explicit installation of DeBEIR
