### 2022-12-15 Ayuei/DeBEIR#4 @Ayuei
Thanks for this. I've updated the setup script and dev environment setup instructions.

Should be working now, and also should fail on any errors.
I must've renamed the test files at some point (and also assumed DeBEIR was installed).
