### 2022-12-16 #5017 @KonradHoeffner
@ayuei: The paper has three authors but you are the [only contributor](https://github.com/Ayuei/DeBEIR/graphs/contributors) of the repository. Can you clarify the contribution of your coauthors?
