### 2022-12-16 #5017 @KonradHoeffner
@Ayuei: Is it “DeBIR” or “DeBEIR”? The paper is titled DeBIR but the repository title and README contains “DeBEIR”. The code in the paper uses both. Please use a consistent name. Or are they actually two different things?
