### 2022-12-16 #5017 @KonradHoeffner
> Additional hackability

@Ayuei: I didn’t know the word “hackability” in this context but I’m not a native speaker and had to look it up to find out that it has two different meanings. If possible I would use a less ambiguous term to not confuse other non-native speakers.
