### 2022-12-16 #5017 @KonradHoeffner
* “A brief summary of the pipeline stages are” summary is singular so it should be “is”, not “are”. Also please refer to figure 1 here, which illustrates the stages.
* “Figure” is repeated several times as “Figure Figure”
* “Similar libraries that exist include sentence-transformers, openNIR” add “such as”, cite publication
