### 2022-12-16 Ayuei/DeBEIR#5 @KonradHoeffner
virtualenv vs venv in  installation instructions
```
$ python3 -m virtualenv venv
/usr/bin/python3: No module named virtualenv
```
Either add `pip install virtualenv` to the docs or use `-m venv` instead if that is enough, as it is shipped by default with Python3.3+, see <https://stackoverflow.com/questions/41573587/what-is-the-difference-between-venv-pyvenv-pyenv-virtualenv-virtualenvwrappe>.

Part of the JOSS review <https://github.com/openjournals/joss-reviews/issues/5017>.
