### 2022-12-16 Ayuei/DeBEIR#6 @KonradHoeffner
please add community guidelines
> **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Part of the JOSS review <https://github.com/openjournals/joss-reviews/issues/5017>
