### 2022-12-16 Ayuei/DeBEIR#7 @KonradHoeffner
please add DeBEIR to PyPI
The documentation instructs to install from source, but ideally it should be installable from PyPI.

Part of the JOSS review <https://github.com/openjournals/joss-reviews/issues/5017>
