### 2022-12-16 Ayuei/DeBEIR#8 @KonradHoeffner
HTML documentation starting page about NIR, not DeBEIR
docs/debeir.html:

>  debeir
>
> The NIR (Neural Index Ranker) source code library.
>
> See ./main.py in the parent directory for an out-of-the-box runnable code.
>
> Otherwise, check out notebooks in the parent directory for training your own model amongst other things.
>

Part of the JOSS review <https://github.com/openjournals/joss-reviews/issues/5017>.
