### 2022-12-26 #5017 @arfon
@Ayuei – please shorten your BibTeX file to include only those entries cited by this paper.
