### 2023-01-11 #5017 @Ayuei
@arfon Hi, I've shortened the BibTex and recompiled the paper.

@KonradHoeffner  Thank you for taking time to review the work and for the detailed comments, I'm still working on addressing concerns and issues.
