### 2023-01-16 #5017 @kuutsav
> 👋 @kuutsav would you be willing to review this submission for JOSS? The submission under consideration is [DeBIR: A Python Package for Dense Bi-Encoder Information Retrieval](https://github.com/openjournals/joss-reviews/issues/4920), a Python package for information retrieval and natural language processing.
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged. You can learn more about the process in these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
> Based on your experience, we think you might be able to provide a great review of this submission. Please let me know if you think you can help us out!
>
> Many thanks Arfon

@arfon Thanks for reaching out. I have a very busy schedule this month. If you are still looking for a reviewer at the end of the month, please reach out to me again. All the best!
