### 2023-01-25 b75619f
Directory refactoring of source code, no major breaking changes.
debeir.interfaces -> debeir.core
debeir.data_sets -> debeir.datasets

- Documentation updated with refactoring
- Test cases updated with refactoring
