+++ src/debeir/core/__init__.py
@@ -1,0 +1,5 @@
+"""
+Core library interfaces that must be implemented for custom datasets
+
+Interfaces to implement custom datasets in debeir.datasets.
+"""
