+++ examples/trec2022/training.py
@@ -12,8 +12,8 @@ from training.utils import DatasetToSentTrans
 from training.hparm_tuning.trainer import SentenceTransformerTrainer
 
 from sentence_transformers import evaluation
-from data_sets.trec_clinical_trials import TrecClinicalTrialTripletParser
-from interfaces import ParsedTopicsToDataset
+from datasets.trec_clinical_trials import TrecClinicalTrialTripletParser
+from core import ParsedTopicsToDataset
 from training.hparm_tuning.config import HparamConfig
 
 
