+++ src/debeir/datasets/__init__.py
@@ -1,5 +1,5 @@
 """
-Contains data_sets implemented from nir.interfaces
+Contains data_sets implemented from nir.core
 1. Parser (For reading data from files into a Dict object)
 2. Query object (Generating queries)
     - These query objects can be very lightweight containing only the mappings of the index.
