+++ CONTRIBUTING.md
@@ -1,0 +1,25 @@
+# Contributing to DeBEIR
+
+Thanks for your interest in contributing! :tada:
+
+## How do I report a bug or suggest an enhancement?
+
+- As a first step, please search in the [existing issues](https://github.com/Ayuei/DeBEIR/issues) to check if your point has already been addressed.
+- If that is not the case, go ahead and [create an issue](https://github.com/Ayuei/DeBEIR/issues/new/choose) of the respective type, providing the details as instructed in the template.
+
+## How do I submit a change?
+
+We welcome contributions via pull requests:
+
+- Fork (or clone) the repo and create your branch from the default branch
+- If you have added code that should be tested - add tests
+- If any documentation updates are needed - make them
+- Ensure the test suite passes and the code lints
+- Submit the pull request
+
+Once you have submitted your PR:
+
+- Note that a PR is considered for review only if it doesn't break existing functionality/test cases
+- Upon approval, PR is to be merged using the "squash and merge" option, so that the commit history remains linear and readable
+
+_Adapted from [FuseMedML](/BiomedSciAI/fuse-med-ml) CONTRIBUTING._
