### 2023-01-25 Ayuei/DeBEIR#7 @Ayuei
Unfortunately this is not possible at the moment as there are direct (git) dependencies in the requirements, which is currently not allowed by PyPi. The solution would be first upload the direct dependencies to PyPi and reference them in DeBEIR.

However, these direct dependencies are other packages I am working on which I don't think are ready yet to be publicly uploaded to PyPi.

As a compromise, I'll be adding releases of the built wheel to this repository.
