### 2023-01-25 #5017 @KonradHoeffner
@arfon I need some help deciding about [this criterion](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort):

> JOSS requires that software should be feature-complete (i.e., no half-baked solutions), packaged appropriately according to common community standards for the programming language being used.

According to <https://github.com/Ayuei/DeBEIR/issues/7>, DeBEIR cannot be published on PyPi because of git dependencies but releases of the built wheel will be added to the repository. Is this enough to satisfy the criterion?
