+++ .github/workflows/docs.yml
@@ -15,28 +15,27 @@ permissions:
 
 jobs:
   # Build the documentation and upload the static HTML files as an artifact.
-  #build:
-  #  runs-on: ubuntu-latest
-  #  steps:
-  #    - uses: actions/checkout@v3
-  #    - uses: actions/setup-python@v4
-  #      with:
-  #        python-version: '3.10'
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-python@v4
+        with:
+          python-version: '3.10'
 
       # ADJUST THIS: install all dependencies (including pdoc)
-  #    - run: pip install -r requirements.txt && pip install pdoc
+      - run: pip install -r docs/requirements.txt && pip install pdoc
       # ADJUST THIS: build your documentation into docs/.
       # We use a custom build script for pdoc itself, ideally you just run `pdoc -o docs/ ...` here.
-  #    - run: pdoc -o docs/ src/debeir/
-
-  #    - uses: actions/upload-pages-artifact@v1
-  #      with:
-  #        path: docs/
+      - run: pdoc -o docs/ src/debeir/
+      - uses: actions/upload-pages-artifact@v1
+        with:
+          path: docs/
 
   # Deploy the artifact to GitHub pages.
   # This is a separate job so that only actions/deploy-pages has the necessary permissions.
   deploy:
-    #needs: build
+    needs: build
     runs-on: ubuntu-latest
     permissions:
       pages: write
