+++ .github/workflows/docs.yml
@@ -24,7 +24,7 @@ jobs:
           python-version: '3.10'
 
       # ADJUST THIS: install all dependencies (including pdoc)
-      - run: pip install -r docs/requirements.txt && pip install pdoc
+      - run: pip install -r requirements-docs.txt && pip install pdoc
       # ADJUST THIS: build your documentation into docs/.
       # We use a custom build script for pdoc itself, ideally you just run `pdoc -o docs/ ...` here.
       - run: pdoc -o docs/ src/debeir/
