+++ requirements-docs.txt
@@ -13,7 +13,7 @@ toml~=0.10.2
 dill~=0.3.5.1
 requests~=2.28.0
 loguru~=0.6.0
-https://github.com/Ayuei/analysis_tools/releases/download/python/analysis_tools_ir-0.0.1-py3-none-any.whl
+https://github.com/Ayuei/analysis_tools_ir/releases/download/v0.0.3/analysis_tools_ir-0.0.3-py3-none-any.whl
 git+https://github.com/Ayuei/allRank.git
 transformers==4.22.0
 shutup~=0.2.0
