### 2023-01-25 #5017 @Ayuei
> @Ayuei: Where are the HTML docs published? Please add a link to that to the README. Right now I can only preview it at https://raw.githack.com/Ayuei/DeBEIR/main/docs/debeir.html.

Hi, I set up a GitHub pages here:
https://ayuei.github.io/DeBEIR/debeir.html
