### 2023-01-29 #5017 @arfon
> According to https://github.com/Ayuei/DeBEIR/issues/7, DeBEIR cannot be published on PyPi because of git dependencies but releases of the built wheel will be added to the repository. Is this enough to satisfy the criterion?

@KonradHoeffner – yes, I think this is acceptable. Thanks for checking.
