+++ paper.md
@@ -3,67 +3,71 @@ title: 'DeBIR: A Python Package for Dense Bi-Encoder Information Retrieval'
 tags:
   - information retrieval
   - dense retrieval
-  - bi-encoder 
+  - bi-encoder
   - transformers
   - pytorch
   - python
   - deep learning
   - neural networks
   - machine learning
-  - natural language processing 
+  - natural language processing
 authors:
   - name: Vincent Nguyen
-    orcid: 0000-0003-1787-8090 
+    orcid: 0000-0003-1787-8090
     affiliation: "1, 2"
-  - name: Sarvnaz Karimi 
-    orcid: 0000-0002-4927-3937 
+  - name: Sarvnaz Karimi
+    orcid: 0000-0002-4927-3937
     affiliation: "2"
-  - name: Zhenchang Xing 
-    orcid: 0000-0001-7663-1421 
+  - name: Zhenchang Xing
+    orcid: 0000-0001-7663-1421
     affiliation: "1, 2"
 affiliations:
- - name: Australian National University 
+ - name: "Australian National University, School of Computing"
    index: 1
- - name: "CSIRO's Data61"
+ - name: "Commonwealth Scientific and Industrial Research Organisation, Data61"
    index: 2
-date: 4 October 2022 
+date: 4 October 2022
 bibliography: paper.bib
 ---
 
 # Summary
-Information Retrieval (IR) is the task of retrieving documents given a query or information need. These documents are retrieved and ranked based on a relevance function or relevance model such as Best-Matching 25 [@bm25]. Although deep learning has been successful in other computer science fields such as computer vision [@alexnet; @googlenet] and natural language processing [@orig-bert-2018; @liu:2019; @biobert], success in information retrieval was limited until 2021 [@lin-neural-recantation]. In 2021, deep learning in information retrieval could surpass less computationally intensive keyword-based statistical models in terms of retrieval effectiveness [@yang2019critically], sparking the field of dense retrieval. Dense retrieval is the task of retrieving documents given a query or information need using a dense vector representation of the query and documents. The dense vector representation is obtained by passing the query and documents through a neural network. The neural network is usually a pre-trained language model such as BERT [@orig-bert-2018] or RoBERTa [@roberta]. The dense query vector representation is then used to retrieve documents using a similarity function such as cosine similarity.
+Information Retrieval (IR) is the task of retrieving documents given a query or information need. These documents are retrieved and ranked based on a relevance function or relevance model such as Best-Matching 25 [@bm25]. Although deep learning has been successful in other computer science fields, such as computer vision [@alexnet; @googlenet] and natural language processing [@orig-bert-2018; @liu:2019; @biobert], success in information retrieval was limited in the literature due to comparisons against weak baselines [@yang2019critically]. However, in 2019 [@lin-neural-recentation], deep learning in information retrieval could surpass less computationally intensive keyword-based statistical models in terms of retrieval effectiveness, sparking the field of dense retrieval. Dense retrieval is the task of retrieving documents given a query or information need using a dense vector representation of the query and documents. The dense vector representation is obtained by passing the query and documents through a neural network. The neural network is usually a pre-trained language model such as BERT [@orig-bert-2018] or RoBERTa [@roberta]. The dense query vector representation is then used to retrieve documents using a similarity function such as cosine similarity.
 
-Unlike statistical learning, tuning of deep learning retrieval methods is often costly and time-consuming. This makes it important to efficiently automate much of the training, tuning and evaluation processes. 
+Unlike statistical learning, tuning deep learning retrieval methods is often costly and time-consuming. This cost makes it essential to automate much of the training, tuning and evaluation processes efficiently.
 
[…]
