+++ paper.md
@@ -35,7 +35,7 @@ Information Retrieval (IR) is the task of retrieving documents given a query or 
 
 Unlike statistical learning, tuning deep learning retrieval methods is often costly and time-consuming. This cost makes it essential to automate much of the training, tuning and evaluation processes efficiently.
 
-DeBIR is a library for facilitating dense retrieval research, primarily focusing on bi-encoder dense retrieval where query and documents dense vectors are generated separately [@reimers2019]. It allows for expedited experimentation in dense retrieval research by reducing boilerplate code through an interchangeable pipeline API and code extendability through the inheritance of general classes. It further abstracts standard training loops and hyperparameter tuning into easy-to-define configuration files. This library is aimed at helping practitioners, researchers and data scientists experimenting with bi-encoders by providing them with dense retrieval methods that are easy to use out of the box but also have additional extendability for more nuanced research.
+DeBIR is a library for facilitating dense retrieval research, primarily focusing on bi-encoder dense retrieval where query and documents dense vectors are generated separately [@reimers2019]. It allows for expedited experimentation in dense retrieval research by reducing boilerplate code through an interchangeable pipeline API and code extendability through the inheritance of general classes. It further abstracts standard training loops and hyperparameter tuning into easy-to-define configuration files. This library is aimed at helping practitioners, researchers and data scientists experimenting with bi-encoders by providing them with dense retrieval methods that are easy to use out of the box but also have additional extendability for more nuanced research. Furthermore, our pipeline runs asynchronously to reduce I/O performance bottlenecks, facilitating faster experiments and research.
 
 A brief summary of the pipeline stages (\autoref{fig:training}) is:
 
@@ -65,9 +65,9 @@ DeBIR is a library that mainly facilitates bi-encoder research (where query and 
 
 The DeBIR library exposes an API for commonly used functions for training, hyper-parameter tuning (\autoref{fig:training}) and evaluation of transformer-based models. The pipeline can be broken up into multiple stages: parsing, query building, query execution, serialization and evaluation (\autoref{fig:pipeline}). Furthermore, we package our caching mechanism for the expensive encoding operations to speed up the pipeline during repeated experimentation.
 
-Although similar libraries exist, such as sentence-transformers [@reimers2019], openNIR[@opennir], they have less of a focus on the early stages of the dense retrieval pipeline. This stage involves indexing the textual data from the corpora and indexing dense vector representations, which is only helpful for bi-encoder type models over the traditional cross-encoder. Other limitations include a lack of extendability that restrict the users' options for training customization (we provide base classes that can be inherited) or that the library is tailored to general-purpose machine learning rather than informational retrieval.
+Although similar libraries exist, such as sentence-transformers [@reimers2019], openNIR[@opennir], they have less of a focus on the early stages of the dense retrieval pipeline. This stage involves indexing the textual data from the corpora and indexing dense vector representations, which is only helpful for bi-encoder type models over the traditional cross-encoder and is thus not typically explored by other libraries. Other limitations include a lack of extendability that restrict the users' options for training customization (we provide base classes that can be inherited) or that the library is tailored to general-purpose machine learning rather than informational retrieval.
 
-This library will help facilitate early-stage dense retrieval and rapid experimentation research while being flexible enough for second-stage retrieval from either this library or other libraries. Furthermore, we will continue to improve this tool.
+This library will help facilitate early-stage dense retrieval and rapid experimentation research with bi-encoders, which other libraries have yet to explore. Our library is also flexible enough for second-stage retrieval using cross-encoders from this library or other libraries. Furthermore, we will continue to improve this tool.
 
 ![Standard flow of the DeBIR query/evaluation loop.\label{fig:pipeline}](pipeline.pdf){scale=0.5}
 
