### 2023-01-30 #5017 @Ayuei
Thank you once again for taking the time to review the paper.

> @Ayuei: The paper has three authors but you are the [only contributor](https://github.com/Ayuei/DeBEIR/graphs/contributors) of the repository. Can you clarify the contribution of your coauthors?

The co-authors helped with the writing of this paper (Sarvnaz Karimi).
And with the original paper/ that the idea was from and projection supervision and direction (Zhenchang Xing + Sarvaz Karimi).

> @Ayuei: Is it “DeBIR” or “DeBEIR”? The paper is titled DeBIR but the repository title and README contains “DeBEIR”. The code in the paper uses both. Please use a consistent name. Or are they actually two different things?

This is indeed a typo, I've fixed this in the paper.

> @Ayuei: A paper from 2019 is cited next, which talks about how the breakthrough came in 2018 or 2019. Please confirm that 2021 is indeed the correct year.

2021 is the incorrect year, the paper I cited was published in 2021 but was submitted in 2019. This is fixed in the current commit.

> > Additional hackability
>
> @Ayuei: I didn’t know the word “hackability” in this context but I’m not a native speaker and had to look it up to find out that it has two different meanings. If possible I would use a less ambiguous term to not confuse other non-native speakers.

I'll replace "hackability" with extendability, as it's intended to mean that a user can use the building blocks in the library to create their own classes or are able to use the batteries-included pipeline class.

> Summary is singular so it should be “is”, not “are”. Also please refer to figure 1 here, which illustrates the stages.
>  “Figure” is repeated several times as “Figure Figure”
> Add “such as”, cite publication

Fixed in the revised paper [bef8baa](https://github.com/Ayuei/DeBEIR/commit/bef8baa667b0e85fde85857b9fe698cdb048805d)

>>    but have less of a focus on all stages of the dense retrieval pipeline

> Please explain in more detail which stages of the pipeline are not covered enough by all existing libraries and what the benefits of DeBIR over those existing libraries are. This is important to clarify, how exactly it "either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)." as part of the "substantial scholarly effort" criterion and for the statement of need criterion.

I've added more detail to the paper in this commit ([79e70e1](https://github.com/Ayuei/DeBEIR/commit/79e70e1b1f044a3e900a35f5d077e2c7929e9cfb)), which hopefully clarifies this point. The point being made is that a library solely for bi-encoder IR research does not yet exist. Though apart from that, effort was made to make sure the library is as fast (execution-wise and ease-of-use-wise) as possible for experimentation.
