### 2023-01-30 #5017 @KonradHoeffner
> The co-authors helped with the writing of this paper (Sarvnaz Karimi). And with the original paper/ that the idea was from and projection supervision and direction (Zhenchang Xing + Sarvaz Karimi).

If both did specific project supervision and direction the criteria seem to be fulfilled and I will mark the appropriate checkbox, thanks for clarifying!
