### 2023-01-30 #5017 @KonradHoeffner
Community guidelines exist now, checked the box.
@Ayuei please convert your reference to ./CONTRIBUTING.md in the readme file into a link.
Using GitHub Markdown syntax, `<./CONTRIBUTING.md> or something similar should work with relative links to files in the repository.
