### 2023-01-30 #5017 @KonradHoeffner
> > @Ayuei: Is it “DeBIR” or “DeBEIR”? The paper is titled DeBIR but the repository title and README contains “DeBEIR”. The code in the paper uses both. Please use a consistent name. Or are they actually two different things?
>
> This is indeed a typo, I've fixed this in the paper.

I just regenerated the paper but it still has many occurrences of both "debir" and "debeir".
