### 2023-01-30 7fb19fb
Fixed occurence of DeBIR, and included new bibtex
