### 2023-01-30 #5017 @Ayuei
> > > @Ayuei: Is it “DeBIR” or “DeBEIR”? The paper is titled DeBIR but the repository title and README contains “DeBEIR”. The code in the paper uses both. Please use a consistent name. Or are they actually two different things?
> >
> >
> > This is indeed a typo, I've fixed this in the paper.
>
> I just regenerated the paper but it still has many occurrences of both "debir" and "debeir".

Sorry, I thought the typos were only in the code examples. Fixed throughout the paper now.

> The reference "openNIR(opennir?)" on line 83 is displayed incorrectly.

Forgot to commit the bibtex, added it now.

> Community guidelines exist now, checked the box. @Ayuei please convert your reference to ./CONTRIBUTING.md in the readme file into a link. Using GitHub Markdown syntax, `<./CONTRIBUTING.md> or something similar should work with relative links to files in the repository.

Added to the repository. Thanks!
