### 2023-01-30 #5017 @KonradHoeffner
There are still spaces in front of full stops: "DeBeIR .", "issues .". Also "github" *-> "GitHub". Hugggingface -> Huggingface.
