### 2023-02-02 6c6761a
Fixed grammatical errors, writing, linking and space issues
