### 2023-02-02 #5017 @Ayuei
@KonradHoeffner thank you again for your time.

> There are still spaces in front of full stops: "DeBeIR .", "issues .". Also "github" *-> "GitHub". Hugggingface -> Huggingface.
Fixed. There was an issue with links in the final pdf which caused the spacing issue.

> Reference lin-neural-recentation is broken.
Fixed this typo.

Additionally, fixed grammar mistakes, and refined explanations in the text. Currently working on refining the documentation.
