### 2023-02-02 #5017 @KonradHoeffner
There are still writing issues in the paper, for example:
* "openNIR(": missing space
* accurate/effective": use a word instead of a slash or choose one
* "encoded at ranking time) (Lin, 2019)": I would include the citation in the parenthesis and not open another set of parentheses there
* "opening a thread in DeBEIR issue": I think creating an issue sounds better than opening a thread in issues.
* `the" neural hype"` in references
