### 2023-02-03 f706b61
Added mandatory cleanup pre-script.
