+++ src/debeir/training/losses/__init__.py
@@ -1,0 +1,6 @@
+"""
+This library contains typical loss functions for training/fine-tuning models to have a notion of relevance.
+
+Typically, off-the-shelf models and universal sentence encoders can typically find duplicates well or inference.
+But relevance is not equivalent to inference and such requires a different loss function for training.
+"""
