+++ src/debeir/engines/dummyindex/__init__.py
@@ -1,0 +1,3 @@
+"""
+A dummy test index. This is an in-memory index.
+"""
