+++ src/debeir/rankers/reranking/__init__.py
@@ -1,0 +1,6 @@
+"""
+This library exposes re-ranking suitable for second-stage retrieval.
+For instance, a bi-encoder could be used for second-stage retrieval after BM25.
+
+Or, a bi-encoder is used for first-stage retrieval followed by a cross-enccoder retrieval.
+"""
