+++ src/debeir/training/hparm_tuning/optuna_rank.py
@@ -1,13 +1,29 @@
+"""
+Contains utility hyperparameter tuning methods using optuna with integration with wandb
+"""
+
 from functools import partial
 
 import joblib
 import optuna
-from debeir.training.hparm_tuning.trainer import Trainer
 from optuna.integration.wandb import WeightsAndBiasesCallback
 from optuna.trial import TrialState
 
+from debeir.training.hparm_tuning.trainer import Trainer
+
 
 def objective(trainer: Trainer, trial: optuna.Trial):
+    """
+    The objective to optimise. This can and should be overriden by a user's implementation.
+
+    :param trainer:
+    :type trainer:
+    :param trial:
+    :type trial:
+    :return:
+    :rtype:
+    """
+
     dataset = trainer.dataset_loading_fn()
 
     train_dataset = dataset["train"]
@@ -20,9 +36,12 @@ def run_optuna_with_wandb(trainer, n_trials=100, n_jobs=1, maximize_objective=Tr
                           wandb_kwargs=None):
     """
     Partially initialize the objective function with a trainer and hparams to optimize.
-
     Optimize using the optuna library.
 
+    :param n_jobs:
+    :type n_jobs:
+    :param save_study_path:
+    :type save_study_path:
     :param trainer:
     :param n_trials:
     :param maximize_objective:
@@ -49,6 +68,14 @@ def run_optuna_with_wandb(trainer, n_trials=100, n_jobs=1, maximize_objective=Tr
 
 
 def print_optuna_stats(study: optuna.Study):
+    """
+    Print the statistics (best run, num completed, num pruned, num trials) from an optuna study
+
+    :param study:
+    :type study:
+    :return:
+    :rtype:
+    """
     pruned_trials = study.get_trials(deepcopy=False,
                                      states=[TrialState.PRUNED])
     complete_trials = study.get_trials(deepcopy=False,
