### 2023-02-06 ef561e0
Added pipeline example
