+++ README.md
@@ -40,6 +40,13 @@ API Documentation for the library with rendered HTML documentation is available 
 
 Statically compiled documentation (which is updated less frequently) can be found in the top level directory [docs/index.html](docs/index.html).
 
+You can also build this documentation with the pdoc3 library by executing the following commands:
+```
+pip install -r requirements-docs.txt
+
+pdoc -o docs/ src/debeir/
+```
+
 ### Development
 
 If you use to help with development of the library, first verify the tests cases and set up a development environment. 
