+++ paper.bib
@@ -190,3 +190,15 @@ keywords = {reproducibility, ad-hoc ranking, neural ranking},
 location = {Houston, TX, USA},
 series = {WSDM '20}
 }
+
+@article{lin2021pretrained,
+  title={Pretrained transformers for text ranking: Bert and beyond},
+  author={Lin, Jimmy and Nogueira, Rodrigo and Yates, Andrew},
+  journal={Synthesis Lectures on Human Language Technologies},
+  volume={14},
+  number={4},
+  pages={1--325},
+  year={2021},
+  publisher={Morgan \& Claypool Publishers}
+}
+
