+++ paper.md
@@ -53,19 +53,18 @@ A brief summary of the pipeline is (\autoref{fig:training}):
 This pipeline is condensed into a single class that can be built from a configuration file.
 
 # Statement of Need
-Dense retrieval has been popular in Information Retrieval since 2015 [@paccr; @drmm; @abcnn]. In the early 2000s, there had been considerable stagnation in retrieval effectiveness as there needed to be more robust baselines [@Armstrong:2009] when proposing new methods. This stagnation repeated with the rise of deep learning, where retrieval performance was again compared against weaker baselines and was not significantly stronger than older non-deep learning statistical models, such as a well-tuned BM25 model [@yang2019critically].
+Dense retrieval has been popular in Information Retrieval since 2015 [@paccr; @drmm; @abcnn]. In the early 2000s, there was a slow down in retrieval effectiveness as there needed to be more robust baselines [@Armstrong:2009] when proposing new methods. This stagnation repeated with the rise of deep learning, where retrieval effectiveness was again compared against weaker baselines and was not shown significantly stronger than statistical models [@yang2019critically], such as a well-tuned BM25 model.
 
-However, this was later recanted when transformer models could be used fine-tuned on Natural Language Inference tasks or ms-marco as a cross-encoder (where a query and document pair are encoded at ranking time) [@lin-neural-recantation], significantly overtaking even the best BM25 models.
+However, this was later recanted when transformer models could be used fine-tuned on Natural Language Inference tasks or Ms-Marco [@msmarco] as a cross-encoder [@lin-neural-recantation], significantly overtaking even the best BM25 models.
 
-Today, there are generally two classes of dense retrieval models for IR: the cross-encoder, which encodes queries and documents at query time and the bi-encoder, which can encode documents at index time and queries at query time. The cross-encoder is generally more effective than the bi-encoder model for retrieval. However, this increased effectiveness requires a more substantial computation and can be a bottleneck in production systems. Therefore, a less expensive model such as BM25 is typically used to retrieve smaller candidate lists (first-stage retrieval) for second-stage retrieval re-ranking by a cross-encoder.
+There are generally two classes of dense retrieval models for IR: (1) the cross-encoder, which encodes queries and documents at query time and (2) the bi-encoder, which can encode documents at index time and queries at query time. The cross-encoder is generally more effective than the bi-encoder model for retrieval [@lin2021pretrained]. However, this increased effectiveness requires a more substantial computation and can be a bottleneck in production systems. Therefore, a less expensive model such as BM25 is typically used to retrieve smaller candidate lists (first-stage retrieval) to be fed to second-stage retrieval re-ranking by a cross-encoder.
 
-Although cross-encoders are more accurate/effective than bi-encoders, the utility of the bi-encoder is that they are more effective than BM25 [@search-like-an-expert-2022] and are faster than cross-encoders. Therefore, a gap in the literature in IR is to replace BM25 first-stage retrieval with a bi-encoder or
-used as the only ranking system in the pipeline if query speed is needed.
-However, current libraries don't address this use case as it requires integration with the indexing and querying pipeline of the search engine.
+Although cross-encoders are more accurate than bi-encoders, bi-encoder are more effective than BM25 [@search-like-an-expert-2022] and are faster than cross-encoders. Therefore, a gap in the literature in IR is to replace BM25 first-stage retrieval with a bi-encoder or
+use as the only ranking system, without a second-stage re-ranker. However, current libraries do not address this use case because it requires integration with the indexing and querying pipeline of the search engine.
 
[…]
