### 2023-02-28 #5017 @KonradHoeffner
> Furthermore, our pipeline runs asynchronously to reduce I/O performance bottlenecks, facilitating faster experiments and research

To check the "performance" box in the review I need some kind of proof for that statement, do you have a benchmark or some kind of measurement to confirm that it is faster than the competitors?
