### 2023-03-30 3f9ede2
Added Lazy Static utility class. Allows for re-use of functions and objects, but also they are lazily executed
