### 2023-03-30 592f03d
Added README.md and symlinks for semantic indexing usage
