### 2023-03-30 bcfe6b0
Elaborated on instructions for semantic indexing
