+++ benchmark/nir.toml
@@ -1,0 +1,28 @@
+[metrics]
+[metrics.common]
+metrics = ["ndcg@10", "ndcg@20", "ndcg@100",
+    "rprec@1000",
+    "p@1", "p@5", "p@10", "p@15", "p@20",
+    "bpref@1000",
+    "recall@1000",
+    "rprec@1000",
+    "r@1000"]
+
+[search.engines]
+[search.engines.elasticsearch]
+protocol = "http"
+ip = "localhost"
+port = "9200"
+timeout = 600
+
+[search.engines.solr]
+ip = "127.0.0.1"
+port = "9200"
+
+
+[nir]
+[nir.default_settings]
+norm_weight = "2.15"
+evaluate = false
+return_size = 10
+output_directory = "./output/test/"
