### 2023-03-30 c6bd9a4
Added a synchronous elasticsearch for benchmarking purposes.
