+++ requirements.txt
@@ -1,28 +1,27 @@
 #torch==1.12.0
-elasticsearch[async]==8.3.1
-plac~=1.3.5
-sentence-transformers==2.2.2
-spacy~=3.2.4
+elasticsearch[async]>=8.3.1
+plac>=1.3.5
+sentence-transformers>=2.2.2
+spacy>=3.2.4
 scispacy
 https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.0/en_core_sci_md-0.5.0.tar.gz
-tqdm~=4.64.0
-pandas~=1.5.3
+tqdm>=4.64.0
+pandas>=1.5.3
 trectools
-numpy~=1.22.4
-toml~=0.10.2
-dill~=0.3.5.1
-requests~=2.28.0
-loguru~=0.6.0
+numpy>=1.22.4
+toml>=0.10.2
+dill>=0.3.5.1
+requests>=2.28.0
+loguru>=0.6.0
 https://github.com/Ayuei/analysis_tools_ir/releases/download/v0.0.3/analysis_tools_ir-0.0.3-py3-none-any.whl
 git+https://github.com/Ayuei/allRank.git
-transformers==4.22.0
-shutup~=0.2.0
-datasets==2.4.0
-jupyterlab==3.4.7
-scikit-learn==1.1.2
+transformers>=4.22.0
+shutup>=0.2.0
+datasets>=2.4.0
+scikit-learn>=1.1.2
 scipy~=1.8.1
-optuna==3.0.2
-wandb==0.13.3
+optuna>=3.0.2
+wandb>=0.13.3
 torch_optimizer
-joblib~=1.2.0
-torch~=1.12.1
+joblib>=1.2.0
+torch>=2.0.0
