+++ benchmark/sequential.py
@@ -1,22 +1,28 @@
-import shutil
-import tarfile
+import time
+
+from loguru import logger
+from tqdm import tqdm
 
 from debeir import Client, NIRPipeline
 
+logger.disable("debeir")
+
 
 def run_all_queries(client, p):
-    for topic_num in p.engine.query.topics:
-        body = p.engine.query.generate_query_embedding(topic_num, cosine_offset=100000)
+    tasks = []
 
-        res = client.search(
-            index=p.engine.index_name, body=body, size=100
+    for topic_num in p.engine.query.topics:
+        body = p.engine.query.generate_query_embedding(topic_num)
+        tasks.append(
+            {"index": p.engine.index_name, "body": body, "size": 10}
         )
+
+    for task in tqdm(tasks):
+        client.search(**task)
 
 
 if __name__ == "__main__":
-    shutil.copy("../../tests/test_set.tar.gz", ".")
-    with tarfile.open("./test_set.tar.gz", mode="r") as f:
-        f.extractall()
+    tracker = []
 
     p = NIRPipeline.build_from_config(config_fp="./config.toml",
                                       engine="elasticsearch",
@@ -24,4 +30,8 @@ if __name__ == "__main__":
 
     client = Client.build_from_config("elasticsearch_sync", p.engine_config)
 
+    start = time.time()
     run_all_queries(client.es_client, p)
+    end = time.time()
+
+    print(end - start)
