+++ benchmark/warm_start_cache.py
@@ -1,4 +1,7 @@
+import time
+
 from loguru import logger
+from tqdm import tqdm
 
 from debeir import NIRPipeline
 
@@ -6,7 +9,7 @@ logger.disable("debeir")
 
 
 def run_all_queries(p):
-    for topic_num in p.engine.query.topics:
+    for topic_num in tqdm(p.engine.query.topics):
         p.engine.query.generate_query_embedding(topic_num)
 
 
@@ -15,4 +18,8 @@ if __name__ == "__main__":
                                       engine="elasticsearch",
                                       nir_config_fp="./nir.toml")
 
+    start = time.time()
     run_all_queries(p)
+    end = time.time()
+
+    print(end - start)
