### 2023-04-02 #5017 @Ayuei
> @Ayuei To check the "performance" box in the review I need some kind of proof for that statement, do you have a benchmark or some kind of measurement to confirm that it is faster than the competitors?

I added a benchmark to justify the statement that our async pipeline is much faster than a synchronous one [Link to benchmark](https://github.com/Ayuei/DeBEIR/tree/main/benchmark). In summary, it's a 9-10x speedup.

Otherwise, all expensive embedding calls are immediately cached (so subsequent calls are I/O bound rather than GPU bound) which is what is tested in ```test_pipeline:test_query_encoder_cache``` in (test_pipeline.py)[https://github.com/Ayuei/DeBEIR/blob/main/tests/test_pipeline.py] and also benchmarked in [Link to benchmark](https://github.com/Ayuei/DeBEIR/tree/main/benchmark). We believe that this is valuable for researchers as experiments are often repeated several times.
