### 2023-04-03 970121d
Lazy static bugfix: added getter
