+++ examples/hparam_tuning/hparam_tuning_from_config.py
@@ -1,6 +1,7 @@
 # coding: utf-8
 
 import datasets
+import plac
 from sentence_transformers import evaluation
 
 from debeir.training.hparm_tuning.config import HparamConfig
@@ -22,7 +23,7 @@ def remap_label(ex):
     return ex
 
 
-def load_dataset():
+def load_dataset(limit):
     """
     Load and preprocess the SNLI dataset
 
@@ -34,11 +35,16 @@ def load_dataset():
 
     # Use our sentence transformer dataset adapter pattern, allows for huggingface datasets to be used with
     # Sentence transformer API
-    train = SentDataset(dataset['train'].map(remap_label).filter(lambda k: k['label'] != -1),
-                        text_cols=['premise', 'hypothesis'],
-                        label_col='label')
+    select_range = limit if limit != -1 else len(dataset['train'])
 
-    val = SentDataset(dataset['test'].map(remap_label).filter(lambda k: k['label'] != -1),
+    train = SentDataset(
+        dataset['train'].select(range(select_range)).map(remap_label).filter(lambda k: k['label'] != -1),
+        text_cols=['premise', 'hypothesis'],
+        label_col='label')
+
+    select_range = limit if limit != -1 else len(dataset['test'])
+
+    val = SentDataset(dataset['test'].select(range(select_range)).map(remap_label).filter(lambda k: k['label'] != -1),
                       text_cols=['premise', 'hypothesis'],
                       label_col='label')
 
@@ -46,17 +52,27 @@ def load_dataset():
     return {'train': train, 'val': val}
 
 
-if __name__ == "__main__":
+@plac.opt('limit', "Only sample a subset for training", type=int)
+def main(limit=-1):
+    if limit != -1:
+        print(f"Limiting training and validation examples to {limit}")
+    else:
+        print(f"Running full test sample of snli, using the flag --limit [number] to limit the training instances.")
+
     hparam_config = HparamConfig.from_json(
         "hparam_cfg.json"
     )
 
     trainer = SentenceTransformerHparamTrainer(
         # Reuse the dataset with lazy static, so we don't have to do the preprocessing repeatedly
-        dataset_loading_fn=lazy_static("hparam_dataloader", load_dataset),
+        dataset_loading_fn=lazy_static("hparam_dataloader", load_dataset, limit),
         evaluator_fn=evaluation.BinaryClassificationEvaluator,
         hparams_config=hparam_config,
     )
 
     study = run_optuna_with_wandb(trainer)
     print_optuna_stats(study)
+
+
+if __name__ == "__main__":
+    plac.call(main)
