### 2023-04-04 b152a2c
Removed -rm flag in docker, indexing is expensive.
