### 2023-04-04 4fba85e
Added README.md for pipeline.py example
