### 2023-04-04 fa22d16
Added a prehook in the training loop
