### 2023-04-04 fb69fdf
Updated hook method signature
