### 2023-04-04 #5017 @KonradHoeffner
@arfon @Ayuei: I'm finished with my review now and checked all the boxes. However I did not know about the field of dense retrieval before, so the second reviewer should make sure to separately check especially whether all functional checks have been confirmed because I can only verify them to my limited understanding of the subject.
