### 2023-06-24 #5017 @amitkumarj441
I am back to reviewing the codebase and was able to run the python tests without a hitch, given this library has support enable for both ES and Solo which works under a standalone manner. However, the automated tests pipeline via Docker runs on MacOs which not platform-agnostic, though, you may add a bit about instructions on how to run this [dockerized image](https://github.com/Ayuei/DeBEIR/blob/ffc86e1d3bd25fce4774f72f4c6537ba3a941f02/tests/build_test_env.sh#L9) for indexing on Linux. The main pros of this library for reproducing a similar case in IR is the support of extended list of metrics (NDCG@k, p@k, r@k, etc.). Apart from this, the documentation should clarify on the timeout for ES such as in [L16](https://github.com/Ayuei/DeBEIR/blob/ffc86e1d3bd25fce4774f72f4c6537ba3a941f02/tests/nir_config.toml#L16C5-L16C18).
