### 2023-06-27 9efca4c
Added elasticsearch config documentation
