+++ configs/nir.toml
@@ -13,7 +13,7 @@
     protocol = "http"
     ip = "127.0.0.1"
     port = "9200"
-    timeout = 600
+    timeout = 600 # 10 minute request timeout for expensive dense retrieval
 
     [search.engines.solr]
     ip = "127.0.0.1"
