### 2023-06-27 #5017 @Ayuei
> I am back to reviewing the codebase and was able to run the Python tests without a hitch, given this library has support enabled for both ES and Solr which works in a standalone manner. However, the automated tests pipeline via Docker runs on MacOs which is not platform-agnostic, though, you may add a bit about instructions on how to run this [dockerized image](https://github.com/Ayuei/DeBEIR/blob/ffc86e1d3bd25fce4774f72f4c6537ba3a941f02/tests/build_test_env.sh#L9) for indexing on Linux. The main pros of this library for reproducing a similar case in IR is the support of an extended list of metrics (NDCG@k, p@k, r@k, etc.). Apart from this, the documentation should clarify on the timeout for ES such as [L16](https://github.com/Ayuei/DeBEIR/blob/ffc86e1d3bd25fce4774f72f4c6537ba3a941f02/tests/nir_config.toml#L16C5-L16C18).

Thank you for your review @amitkumarj441,

I added more documentation about the ES config in commit [b25dd9](https://github.com/Ayuei/DeBEIR/commit/b25dd9e24be855277f7460b97e42e3fb49f87d0d)

I think the testing should be platform agnostic on the main branch (https://github.com/Ayuei/DeBEIR/blob/main/tests/build_test_env.sh), which has seen major refactoring since the paper submission.
