### 2023-06-28 #5017 @amitkumarj441
@Ayuei That's helpful to run tests, however, under the main branch in the central instructions file (README.md), you should fix the typo of the shell script file as below
```
cd tests/
./build_test.env.sh -> ./build_test_env.sh
pytest .
```
Also, fix the missing and erroneous references as listed above.
I am happy to recommend the paper for publication at this stage provided that the above suggestions are taken into consideration.
