### 2023-06-29 ee2e421
Update paper.bib
Added missing DOIs
Tidied up entries
