+++ paper.bib
@@ -5,7 +5,7 @@
 	month        = {01},
 	booktitle    = {TREC},
 	address      = {Gaithersburg, MD, US},
-	url          = {https://trec.nist.gov/pubs/trec3/t3\_proceedings.html}
+	url          = {https://trec.nist.gov/pubs/trec3/t3\_proceedings.html},
 }
 @incollection{alexnet,
 	title        = {ImageNet Classification with Deep Convolutional Neural Networks},
@@ -14,7 +14,8 @@
 	booktitle    = {Advances in Neural Information Processing Systems 25},
 	publisher    = {Curran Associates, Inc.},
 	pages        = {1097--1105},
-	editor       = {F. Pereira and C. J. C. Burges and L. Bottou and K. Q. Weinberger}
+	doi          = {10.1145/3065386},
+	editor       = {F. Pereira and C. J. C. Burges and L. Bottou and K. Q. Weinberger},
 }
 @inproceedings{googlenet,
 	title        = {Going Deeper with Convolutions},
@@ -26,7 +27,7 @@
 	pages        = {1--9},
 	doi          = {10.1109/CVPR.2015.7298594},
 	issn         = {1063-6919},
-	note         = {arXiv:1409.4842}
+	note         = {arXiv:1409.4842},
 }
 @inproceedings{orig-bert-2018,
 	title        = {{BERT}: Pre-training of Deep Bidirectional Transformers for Language Understanding},
@@ -37,7 +38,7 @@
 	address      = {Minneapolis, MN},
 	pages        = {4171--4186},
 	doi          = {10.18653/v1/N19-1423},
-	url          = {https://www.aclweb.org/anthology/N19-1423}
+	url          = {https://www.aclweb.org/anthology/N19-1423},
 }
 @inproceedings{liu:2019,
 	title        = {Text Summarization with Pretrained Encoders},
@@ -48,7 +49,7 @@
 	address      = {Hong Kong, China},
 	pages        = {3730--3740},
 	doi          = {10.18653/v1/D19-1387},
-	url          = {https://www.aclweb.org/anthology/D19-1387}
+	url          = {https://www.aclweb.org/anthology/D19-1387},
 }
 @article{biobert,
 	title        = {{BioBERT: a pre-trained biomedical language representation model for biomedical text mining}},
@@ -62,7 +63,7 @@
 	doi          = {10.1093/bioinformatics/btz682},
 	issn         = {1367-4803},
 	url          = {https://doi.org/10.1093/bioinformatics/btz682},
-	eprint       = {http://oup.prod.sis.lan/bioinformatics/advance-article-pdf/doi/10.1093/bioinformatics/btz682/30132027/btz682.pdf}
+	eprint       = {http://oup.prod.sis.lan/bioinformatics/advance-article-pdf/doi/10.1093/bioinformatics/btz682/30132027/btz682.pdf},
 }
 @article{lin-neural-recantation,
 	title        = {Neural Hype, Justified! {A} recantation},
@@ -71,7 +72,7 @@
 	journal      = {ACM SIGIR Forum},
 	volume       = 53,
 	url          = {http://sigir.org/wp-content/uploads/2019/december/p088.pdf},
-	issue        = 2
+	issue        = 2,
 }
 @inproceedings{yang2019critically,
 	title        = {Critically Examining the" Neural Hype" Weak Baselines and the Additivity of Effectiveness Gains from Neural Ranking Models},
@@ -80,9 +81,10 @@
 	booktitle    = {Proceedings of the 42nd International ACM SIGIR Conference on Research and Development in Information Retrieval},
 	address      = {Paris, France},
 	pages        = {1129--1132},
[…]
