### 2023-06-29 #5017 @Ayuei
> @Ayuei That's helpful to run tests, however, under the main branch in the central instructions file (README.md), you should fix the typo of the shell script file as below
>
> ```
> cd tests/
> ./build_test.env.sh -> ./build_test_env.sh
> pytest .
> ```
>
> Also, fix the missing and erroneous references as listed above. I am happy to recommend the paper for publication at this stage provided that the above suggestions are taken into consideration.

@amitkumarj441 Thanks for catching this!

I just fixed this in the main and paper branches. I also fixed the missing and erroneous DOIs (Commit https://github.com/Ayuei/DeBEIR/commit/8a9714aed769431aefdd12d1d4eb929cc9682318).
