### 2023-07-01 799e978
Fixed grammar issues and improved writing
