+++ paper.md
@@ -31,7 +31,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-Information Retrieval (IR) is the task of retrieving documents given a query or information need. These documents are retrieved and ranked based on a relevance function or relevance model such as Best-Matching 25 (BM25) [@bm25]. Although deep learning has been successful in other computer science fields, such as computer vision with AlexNet [@alexnet] and Inception [@googlenet] and natural language processing with transformers [@orig-bert-2018; @liu:2019; @biobert]; success in information retrieval was limited due to comparisons against weak baselines [@yang2019critically]. However, in 2019 [@lin-neural-recantation], deep learning in information retrieval could surpass less computationally intensive keyword-based statistical models in terms of retrieval effectiveness, sparking the field of dense retrieval. Dense retrieval is the task of retrieving documents given a query or information need using a dense vector representation of the query and documents [@lin2021pretrained]. The dense vector representation is obtained by passing the query and documents through a neural network. The neural network is usually a pre-trained language model such as BERT [@orig-bert-2018] or RoBERTa [@roberta]. The dense query vector representation is then used to retrieve documents using a similarity function such as cosine similarity.
+Information Retrieval (IR) is the task of retrieving documents given a query or information need. These documents are retrieved and ranked based on a relevance function or relevance model such as Best-Matching 25 (BM25) [@bm25]. Although deep learning has been successful in other computer science fields, such as computer vision with AlexNet [@alexnet] and Inception [@googlenet] and natural language processing with transformers [@orig-bert-2018; @liu:2019; @biobert]; success in information retrieval was limited due to comparisons against weak baselines [@yang2019critically]. However, in 2019 [@lin-neural-recantation], deep learning in information retrieval could surpass less computationally intensive keyword-based statistical models in terms of retrieval effectiveness, sparking a resurgence in the field of dense retrieval. Dense retrieval is the task of retrieving documents given a query or information need using a dense vector representation of the query and documents [@lin2021pretrained]. The dense vector representation is obtained by passing the query and documents through a neural network. The neural network is usually a pre-trained language model such as BERT [@orig-bert-2018] or RoBERTa [@roberta]. The dense query vector representation is then used to retrieve documents using a similarity function such as cosine similarity.
 
 Unlike statistical learning, tuning deep learning retrieval methods is often costly and time-consuming. This cost makes it essential to efficently automate much of the training, tuning and evaluation processes.
 
@@ -53,20 +53,20 @@ A brief summary of the pipeline is (\autoref{fig:training}):
 This pipeline is condensed into a single class that can be built from a configuration file.
 
 # Statement of Need
-Dense retrieval has been popular in Information Retrieval since 2015 [@paccr; @drmm; @abcnn]. In the early 2000s, there was a slow down in retrieval effectiveness as there needed to be more robust baselines [@Armstrong:2009] when proposing new methods. This stagnation repeated with the rise of deep learning, where retrieval effectiveness was again compared against weaker baselines and was not shown significantly stronger than statistical models [@yang2019critically], such as a well-tuned BM25 model.
+Dense retrieval has been popular in Information Retrieval since 2015 [@paccr; @drmm; @abcnn]. Retrieval effectiveness of these dense retrieval methods was often compared against weaker baselines and was not shown significantly stronger than statistical models [@yang2019critically], such as a well-tuned BM25 model while being considerably slower. This situation is similar to what happened in the early 2000s, where there was a slow down in retrieval effectiveness from the use of less robust baselines [@Armstrong:2009] when proposing new methods. 
 
-However, this was later recanted when transformer models could be used fine-tuned on Natural Language Inference tasks or Ms-Marco [@msmarco] as a cross-encoder [@lin-neural-recantation], significantly overtaking even the best BM25 models.
[…]
