### 2023-07-01 #5017 @Ayuei
> @Ayuei – At this point could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
>
> * The title of the archive is the same as the JOSS paper title
> * That the authors of the archive are the same as the JOSS paper authors
> * Finally, please make sure you give the paper a final proof read yourself before I do.
>
> I can then move forward with accepting the submission.

Hi @arfon,

Thank you for the detailed instructions

* I published the new release here with all the changes from the reviews: [Latest Release](https://github.com/Ayuei/DeBEIR/releases/tag/v0.1.2)
* I made the archive here with Zenodo here: [10.5281/zenodo.8103783](https://doi.org/10.5281/zenodo.8103783)
* Finally, I also did proof read and uploaded the changes the paper branch: [paper branch](https://github.com/Ayuei/DeBEIR/tree/paper)

If there is anything else I need to do, please let me know.
